requirements.txt
Authlib==1.3.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==2.1.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
urllib3==2.2.1
Werkzeug==3.0.3
